feature/sprint_9-solution-http-api. Декомпозировал код, разбросал классы по файлам, убрал лишнее
